CHANGED: Auctions table will now store item lowguids. CHANGED: Some smaller changes. Some other tables like guildbank items, mail and whatnot already stored item lowguids yet we stored them as fullguid in memory. Not anymore.
DON'T FORGET TO APPLY THE SQL QUERY! DON'T FORGET TO APPLY THE SQL QUERY! DON'T FORGET TO APPLY THE SQL QUERY! DON'T FORGET TO APPLY THE SQL QUERY! DON'T FORGET TO APPLY THE SQL QUERY! DON'T FORGET TO APPLY THE SQL QUERY! DON'T FORGET TO APPLY THE SQL QUERY!

+++ src/arcemu-world/MailSystem.h
@@ -82,7 +82,7 @@ struct MailMessage
 	string subject;
 	string body;
 	uint32 money;
-	vector<uint64> items;
+	vector<uint32> items;
 	uint32 cod;
 	uint32 stationery;
 	uint32 expire_time;
